mejoras

+++ src/app/componentes/header/header.component.ts
@@ -1,5 +1,5 @@
 import { CommonModule, DatePipe } from '@angular/common';
-import { Component } from '@angular/core';
+import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { Router, RouterLink } from '@angular/router';
 import * as moment from 'moment';
@@ -10,6 +10,8 @@ import { SharedModule } from 'src/app/shared/shared.modules';
 import { RestService } from 'src/app/shared/utilitario/rest.service';
 import { UtilsService } from 'src/app/shared/utilitario/util.service';
 import { StorageService } from './../../shared/Service/storage.service';
+import { FechaEvento } from 'src/app/shared/model/FechaEvento.model';
+import { favorito } from 'src/app/shared/model/Favoritos.model';
 declare var iziToast: any;
 declare var $: any;
 @Component({
@@ -20,6 +22,8 @@ declare var $: any;
   styleUrls: ['./header.component.css']
 })
 export class HeaderComponent {
+  @Output() contactChange:EventEmitter<boolean> =new EventEmitter<boolean>();
+
   public id: any;
   public token = localStorage.getItem('token')
   public user: any = undefined;
@@ -33,7 +37,7 @@ export class HeaderComponent {
   public user_data: any = null;
   public imgUrl = '';
   public idUsuario = '';
-  public loadingRecordar =  false;
+  public loadingRecordar = false;
   public loading = false;
   public PasswordCorrecto = false;
   public emailIncorrecto = false;
@@ -43,6 +47,7 @@ export class HeaderComponent {
   public loadingRegistro = false;
   public mensajeEmail = '';
   public mensajePassword = '';
+  public EventosNoRepetidos: any =[];
   public Loginusuario = {
     email: '',
     password: '',
@@ -65,6 +70,8 @@ export class HeaderComponent {
     v_nombre: false,
     v_apellido: false,
     v_password: false,
+    v_politicaPrivacidad: false,
+    v_terminoCondiciones: false,    
   }
 
   public _fechaSrtMaxima: any = "";
@@ -78,11 +85,9 @@ export class HeaderComponent {
     const _current = this.cookieService.get('_currentUser');
     const _parseCurrent = (_current && JSON.parse(this.cookieService.get('_currentUser'))) ?
       JSON.parse(this.cookieService.get('_currentUser')) : null;
-
+    this.user_data= this._clienteService.getCurrentUser();
     this.cliente = undefined;
     this._fechaSrtMaxima = moment(new Date()).format('YYYY-MM-DD');// new Date(new Date().getFullYear(), new Date().getMonth(), new Date().getDate());
-    console.log('----------------------------------------------------------------->');
-    console.log(this._fechaSrtMaxima);
     if (_parseCurrent) {
       this.cliente = _parseCurrent.nombre;
       this.idUsuario = _parseCurrent.sub
@@ -91,32 +96,41 @@ export class HeaderComponent {
       this.cliente = undefined;
     }
   }
-  ngOnInit(): void {
+  async ngOnInit(): Promise<void> {
     //Called after the constructor, initializing input properties, and the first call to ngOnChanges.
     //Add 'implements OnInit' to the class.
-    if (this.storageService.isLoggedIn()) {
-      this.isLoggedIn = true;
-      console.log('logeado')
-      this.verFavoritosUsuario(this.storageService.getUser());
-    }
+    this.user_data= this._clienteService.getCurrentUser();
+      if(this.user_data!=null){
+       
+        this.validarSesion();
+      }
+    
   }
-  consultar_favoritos_guest() {
-    this.rest.get('/listar_favoritos/' + this.user_data.sub).then((response: any) => {
-      this.eventos = response.data;
-    }).catch((error: any) => {
 
+  validarSesion() {
+    this.rest.get('usuario/auth/' + this.user_data.token).then((response: any) => {
+    
+      if(response.ok){
+        this.isLoggedIn = response.ok;
+      } else {        
+        localStorage.clear();
+        sessionStorage.clear();
+        // this._router.navigate(['/']);
+        this._clienteService.logout();
+        this.reloadPage();
+      }
+      
+     
+    }).catch((error: any) => {
       if (error.status == 401) {
         this.util.openSnackBar('Ups! Su Sesion se ha terminado', 'error');
         this._clienteService.logout();
       } else {
         this.util.openSnackBar('Ups! No se ha Agregado el evento en tus favoritos', 'error');
       }
-      /* window.location.reload();
-       localStorage.clear();
-       this._router.navigate(['/']);*/
-      //this._clienteService.logout();
     });
-  }
+  }  
+
   openCart() {
     var clase = $('#modalCarrito').attr('class');
     if (clase == 'ps-panel--sidebar') {
@@ -138,16 +152,17 @@ export class HeaderComponent {
       $('#dropMenu').css({ "left": "0px", "top": "43px", "display": "none" });
     }
   }
-  EliminarFavoritos(idFavorito: any) {
-
+  EliminarFavoritos(id_evento:any,id_fecha:any) {
     var idFavorito = idFavorito;
     if (!this.user_data) {
       this.util.openSnackBar('Ups! Inicia sesión para agregarlo a tus favoritos', 'warning');
     } else {
-      this.rest.delete('/eliminar_favoritos/' + idFavorito).then((response: any) => {
-        if (response.estado == 1) {
-          this.eventos = response.data;
-          this.consultar_favoritos_guest();
+      this.rest.delete('usuario/eliminar_favoritos/'+id_evento+"/"+id_fecha).then((response: any) => {
+        if (response.ok) {
+         
+          this.verFavoritosUsuario(this.user_data.token,2);
+         
+         
         } else {
           this.util.openSnackBar('Ups! Error al Eliminar en Favoritos', 'error');
         }
@@ -161,9 +176,10 @@ export class HeaderComponent {
 
     localStorage.clear();
     sessionStorage.clear();
-    
-   // this._router.navigate(['/']);
+
+    // this._router.navigate(['/']);
     this._clienteService.logout();
+    this.reloadPage();
   }
 
   openSearch() {
@@ -175,50 +191,50 @@ export class HeaderComponent {
     }
   }
   onKeyUpEvent(e: any) {
-    //console.log(e.target.value);
+    ////console.log(e.target.value);
   }
   CloseSearch() {
     this.searchOpen = false;
   }
   searchText(e: any) {
-
+   try {
     if (e == "") {
       this.searchOpen = false;
       this.isloading = true;
     } else {
       this.searchOpen = true;
       this.isloading = true;
-      this.rest.getConsulta("/listar_eventos_filtro_letras/" + e).then((response: any) => {
+      this.rest.getConsulta("eventos/listar_eventos_filtro_letras/" + e).then((response: any) => {
+        //console.log(response.data);
         if (response.message == 'OK') {
-
-          this.dataDesktop = response.data.sort(function (o1: any, o2: any) {
-            var c = new Date(o1.FechaInicio);
-            var d = new Date(o2.FechaInicio);
-            return c > d ? 1 : -1;
-          });
-
+          this.dataDesktop =  response.data;
+          //console.log(this.EventosNoRepetidos);
           this.isloading = false;
         } if (response.message == 'NOK') {
           this.isloading = false;
-        }
-        else {
-          this.util.openSnackBar('Ups! Error al buscar información', 'error');
+         
         }
       })
 
     }
+   } catch (error) {
+    //console.log(error);
+    this.util.openSnackBar('Ups! Error al buscar información', 'error');
+   }
 
 
   }
   search(value: any) {
     this.isloading = true;
     if (value) {
-      this.rest.getConsulta(`/consultar_total_eventos`).then((response: any) => {
+      this.rest.getConsulta("eventos/listar_eventos_filtro_letras/" + value).then((response: any) => {
+        //console.log(response);
         if (response.message == 'OK') {
           if (value == null || value == "") {
             this.data = [];
           } else {
-            this.data = response.data;
+            this.dataDesktop = response.data;
+            //console.log(this.dataDesktop)
           }
         }
       }).catch((error: any) => {
@@ -247,14 +263,18 @@ export class HeaderComponent {
           'password': this.Loginusuario.password
         }
       ).then(logged => {
+        //console.log('--------------->',logged);
         this.storageService.saveUser(logged);
         this.isLoginFailed = false;
         this.isLoggedIn = true;
-        this.Loginusuario.email="";
-        this.Loginusuario.password="";
-        this.reloadPage();
+        this.Loginusuario.email = "";
+        this.Loginusuario.password = "";
+        this.user_data= this._clienteService.getCurrentUser();
+        //console.log('--------------->',this.user_data);
+        this.verFavoritosUsuario(logged,1);
+      
       }).catch((error) => {
-        console.log(error)
+        //console.log(error)
         this.loading = false;
         this.util.openSnackBar('Ups! ' + error.error.message, 'error');
         this.isLoginFailed = false;
@@ -280,16 +300,16 @@ export class HeaderComponent {
     this.PantallaRegistro = true;
     this.PantallaRecordarContrasena = false;
   }
-  enviarLink(){
+  enviarLink() {
     this.loading = true;
-    if(!this.validaciondeRecordarContrasena()){
+    if (!this.validaciondeRecordarContrasena()) {
       this._clienteService.enviarCorreo(
         {
           'email': this.Loginusuario.email
         }
       ).then(logged => {
         this.loading = false;
-        var datos =  logged;
+        var datos = logged;
 
         this.util.openSnackBar("El Email para la recuperacion ha sido enviado", 'success');
         //this.utils.closeAllModals(this.elem)
@@ -307,37 +327,35 @@ export class HeaderComponent {
     }
   }
 
-  validaciondeRecordarContrasena(){
-    if(this.Loginusuario.email==''){
+  validaciondeRecordarContrasena() {
+    if (this.Loginusuario.email == '') {
       this.util.openSnackBar('ups... campo email vacio', 'warning');
       return true;
     }
     return false;
   }
   registrarCuenta() {
-    console.log('registrarCuenta');
+    //console.log('registrarCuenta');
     if (!this.validacionPerfil()) {
-      console.log('validacion');
+      //console.log('validacion');
       try {
         this.loadingRegistro = true;
-        console.log(this.cliente);
         this._clienteService.registrar_cliente_guest(this.Nuevocliente)
           .then(response => {
-            console.log(response);
             this.loadingRegistro = false;
-            this.util.openSnackBar('Upaa! ya tiene Cuenta  , gracias' , 'success');
-            this.PantallaRegistro =  false;
-            this.PantallaLogin =  true;
-          
-              this.Nuevocliente.nombre = '',
-              this.Nuevocliente.apellido= '',
-              this.Nuevocliente.email= '',
-              this.Nuevocliente.genero= '-1',
-              this.Nuevocliente.f_nacimiento= '',
-              this.Nuevocliente.terminoCondiciones= false,
-              this.Nuevocliente.politicaPrivacidad= false,
-              this.Nuevocliente.password= ''
-           
+            this.util.openSnackBar('!Hurra! ya tiene Cuenta  , gracias', 'success');
+            this.PantallaRegistro = false;
+            this.PantallaLogin = true;
+
+            this.Nuevocliente.nombre = '',
+              this.Nuevocliente.apellido = '',
+              this.Nuevocliente.email = '',
+              this.Nuevocliente.genero = '-1',
+              this.Nuevocliente.f_nacimiento = '',
+              this.Nuevocliente.terminoCondiciones = false,
+              this.Nuevocliente.politicaPrivacidad = false,
+              this.Nuevocliente.password = ''
+
 
           }).catch((error) => {
             this.loadingRegistro = false;
@@ -351,7 +369,7 @@ export class HeaderComponent {
         this.loading = false;
         this.util.openSnackBar("Ups.. error al registrar cuenta", 'Vuelva a intentar');
       }
-      
+
     }
   }
   ingresarCuenta() {
@@ -361,7 +379,6 @@ export class HeaderComponent {
     this.PantallaRecordarContrasena = false;
   }
   validaciondeLogin() {
-    console.log(this.Loginusuario);
     this.mensajeEmail = '';
     this.mensajePassword = '';
     if (this.Loginusuario.email == '') {
@@ -404,26 +421,60 @@ export class HeaderComponent {
       this.ErrorNuevoCliente.v_fechaNacimiento = true;
       return true;
     }
+    if (this.Nuevocliente.politicaPrivacidad == false) {
+      this.ErrorNuevoCliente.v_politicaPrivacidad = true;
+      return true;
+    }    
+    if (this.Nuevocliente.terminoCondiciones == false) {
+      this.ErrorNuevoCliente.v_terminoCondiciones = true;
+      return true;
+    }    
     return false;
   }
-  verFavoritosUsuario(arg0: any) {
-
-
+  consultar_favoritos_guest() {
+    debugger;
+    this.rest.get('usuario/consultarfavoritos/' + this.user_data.token).then((response: any) => {
+      this.eventos = response.data;
+    }).catch((error: any) => {
+      if (error.status == 401) {
+        this.util.openSnackBar('Ups! Su Sesion se ha terminado', 'error');
+        this._clienteService.logout();
+      } else {
+        this.util.openSnackBar('Ups! No se ha Agregado el evento en tus favoritos', 'error');
+      }
+    });
+  }
+  verFavoritosUsuario(arg0: any,tipo: any) {   
+    debugger;
     this._clienteService.consultar_favoritos_guest(
-      arg0.sub
-    ).then(logged => {
+      arg0.token
+    ).then(logged => {  
+      
       var datos = JSON.stringify(logged);
       var informacion = JSON.parse(datos)
+      this.eventos = informacion.data;
       var favoritos: any = [];
       informacion.data.forEach(element => {
-        favoritos.push(element.evento._id);
+        var Modelfavoritos = {
+          idEvento :element.ideventos,
+          idfecha :element.fecha,
+          idFavorito: element.idfavoritos
+        }
+        favoritos.push(Modelfavoritos);
       });
       this.storageService.saveFavoritos(favoritos);
+      if(tipo==2){        
+        this.contactChange.emit(true);
+      } else {
+        this.reloadPage();
+      }
+
     }).catch((error) => {
       this.errorMessage = error.error.message;
       this.isLoginFailed = true;
     });
   }
+  
 }
